Move `EventStreamSymbolProvider` into `codegen-core`
